+++ clang/test/Sema/asm-goto.cpp
@@ -45,7 +45,7 @@ BAR:
   return;
 }
 
-int test3(int n)
+int test3(int n) // expected-note 2{{declared here}}
 {
   // expected-error@+2 {{cannot jump from this asm goto statement to one of its possible targets}}
   // expected-error@+1 {{cannot jump from this asm goto statement to one of its possible targets}}
@@ -53,9 +53,11 @@ int test3(int n)
   // expected-note@+3 {{jump bypasses initialization of variable length array}}
   // expected-note@+2 {{possible target of asm goto statement}}
   // expected-note@+1 {{jump enters a statement expression}}
-  return ({int a[n];label_true: 2;});
+  return ({int a[n];label_true: 2;}); // expected-warning {{variable length arrays are a Clang extension}} \
+                                         expected-note {{function parameter 'n' with unknown value cannot be used in a constant expression}}
   // expected-note@+1 {{jump bypasses initialization of variable length array}}
-  int b[n];
+  int b[n]; // expected-warning {{variable length arrays are a Clang extension}} \
+               expected-note {{function parameter 'n' with unknown value cannot be used in a constant expression}}
 // expected-note@+1 {{possible target of asm goto statement}}
 loop:
   return 0;