+++ clang/test/SemaCXX/typeid.cpp
@@ -22,10 +22,13 @@ void g1(X &x) {
   (void)typeid(x); // expected-error{{'typeid' of incomplete type 'X'}}
 }
 
-void h(int i) {
-  char V[i];
+void h(int i) { // expected-note 2{{declared here}}
+  char V[i]; // expected-warning{{variable length arrays are a Clang extension}} \
+                expected-note{{function parameter 'i' with unknown value cannot be used in a constant expression}}
   typeid(V);        // expected-error{{'typeid' of variably modified type 'char[i]'}}
-  typeid(char [i]); // expected-error{{'typeid' of variably modified type 'char[i]'}}
+  typeid(char [i]); // expected-error{{'typeid' of variably modified type 'char[i]'}} \
+                       expected-warning{{variable length arrays are a Clang extension}} \
+                       expected-note{{function parameter 'i' with unknown value cannot be used in a constant expression}}
 }
 
 // expected-note@+1 {{read of object 'typeid(int).name' whose value is not known}}