+++ clang/test/SemaCXX/weak-init.cpp
@@ -5,9 +5,11 @@ extern const int W1 __attribute__((weak)) = 10; // expected-note {{declared here
 static_assert(W1 == 10, ""); // expected-error   {{static assertion expression is not an integral constant expression}}
                              // expected-note@-1 {{initializer of weak variable 'W1' is not considered constant because it may be different at runtime}}
 
-extern const int W2 __attribute__((weak)) = 20;
+extern const int W2 __attribute__((weak)) = 20; // expected-note {{declared here}}
 
-int S2[W2]; // expected-error {{variable length array declaration not allowed at file scope}}
+int S2[W2]; // expected-error {{variable length array declaration not allowed at file scope}} \
+               expected-warning {{variable length arrays are a Clang extension}} \
+               expected-note {{initializer of weak variable 'W2' is not considered constant because it may be different at runtime}}
 
 extern const int W3 __attribute__((weak)) = 30; // expected-note {{declared here}}